[FrameworkBundle][Cache][Serializer][Validator] Move the cache warmers into their own components
