+++ src/Symfony/Bundle/FrameworkBundle/CacheWarmer/SerializerCacheWarmer.php
@@ -11,74 +11,8 @@
 
 namespace Symfony\Bundle\FrameworkBundle\CacheWarmer;
 
-use Symfony\Component\Cache\Adapter\ArrayAdapter;
-use Symfony\Component\Serializer\Mapping\Factory\CacheClassMetadataFactory;
-use Symfony\Component\Serializer\Mapping\Factory\ClassMetadataFactory;
-use Symfony\Component\Serializer\Mapping\Loader\AttributeLoader;
-use Symfony\Component\Serializer\Mapping\Loader\LoaderChain;
-use Symfony\Component\Serializer\Mapping\Loader\LoaderInterface;
-use Symfony\Component\Serializer\Mapping\Loader\XmlFileLoader;
-use Symfony\Component\Serializer\Mapping\Loader\YamlFileLoader;
+use Symfony\Component\Serializer\CacheWarmer\SerializerCacheWarmer as BaseSerializerCacheWarmer;
 
-/**
- * Warms up serializer metadata.
- *
- * @author Titouan Galopin <galopintitouan@gmail.com>
- */
-final class SerializerCacheWarmer extends AbstractPhpFileCacheWarmer
-{
-    /**
-     * @param LoaderInterface[] $loaders      The serializer metadata loaders
-     * @param string            $phpArrayFile The PHP file where metadata are cached
-     */
-    public function __construct(
-        private array $loaders,
-        string $phpArrayFile,
-    ) {
-        parent::__construct($phpArrayFile);
-    }
+trigger_deprecation('symfony/framework-bundle', '8.2', 'The "%s" class is deprecated, use "%s" instead.', 'Symfony\Bundle\FrameworkBundle\CacheWarmer\SerializerCacheWarmer', BaseSerializerCacheWarmer::class);
 
-    protected function doWarmUp(string $cacheDir, ArrayAdapter $arrayAdapter, ?string $buildDir = null): bool
-    {
-        if (!$buildDir) {
-            return false;
-        }
-        if (!$this->loaders) {
-            return true;
-        }
-
-        $metadataFactory = new CacheClassMetadataFactory(new ClassMetadataFactory(new LoaderChain($this->loaders)), $arrayAdapter);
-
-        foreach ($this->extractSupportedLoaders($this->loaders) as $loader) {
-            foreach ($loader->getMappedClasses() as $mappedClass) {
-                try {
-                    $metadataFactory->getMetadataFor($mappedClass);
-                } catch (\Exception $e) {
-                    $this->ignoreAutoloadException($mappedClass, $e);
-                }
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * @param LoaderInterface[] $loaders
-     *
-     * @return list<XmlFileLoader|YamlFileLoader|AttributeLoader>
-     */
-    private function extractSupportedLoaders(array $loaders): array
-    {
-        $supportedLoaders = [];
-
-        foreach ($loaders as $loader) {
-            if (method_exists($loader, 'getMappedClasses')) {
-                $supportedLoaders[] = $loader;
-            } elseif ($loader instanceof LoaderChain) {
-                $supportedLoaders = array_merge($supportedLoaders, $this->extractSupportedLoaders($loader->getLoaders()));
-            }
-        }
-
-        return $supportedLoaders;
-    }
-}
+class_alias(BaseSerializerCacheWarmer::class, 'Symfony\Bundle\FrameworkBundle\CacheWarmer\SerializerCacheWarmer');
