--- src/Symfony/Bundle/FrameworkBundle/Tests/Fixtures/Serialization/Author.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Serialization;
-
-class Author
-{
-    public $eyeColor;
-}
