--- src/Symfony/Bundle/FrameworkBundle/Tests/Fixtures/Serialization/Person.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Serialization;
-
-class Person
-{
-    public $eyeColor;
-}
