--- src/Symfony/Bundle/FrameworkBundle/Tests/Fixtures/Validation/Author.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Validation;
-
-class Author
-{
-    public $eyeColor;
-}
