+++ src/Symfony/Component/Serializer/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Add `SerializerBundle`, which provides the `serializer` configuration and the services previously provided by `FrameworkBundle` under `framework.serializer`
+ * Add `CacheWarmer\SerializerCacheWarmer`, which `FrameworkBundle` used to provide
  * Add the `force_collection` context option to `XmlEncoder` to always decode the given tags as a collection
  * Add `AbstractNormalizer::SKIP_INVALID_ATTRIBUTES` to denormalize the attributes whose value cannot be used as if they were absent from the input
  * Add `AbstractNormalizer::IGNORED_GROUPS` to exclude the attributes belonging to the given groups
