+++ src/Symfony/Component/Serializer/CacheWarmer/SerializerCacheWarmer.php
@@ -1,0 +1,85 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Serializer\CacheWarmer;
+
+use Symfony\Component\Cache\Adapter\ArrayAdapter;
+use Symfony\Component\Cache\CacheWarmer\AbstractPhpFileCacheWarmer;
+use Symfony\Component\Serializer\Mapping\Factory\CacheClassMetadataFactory;
+use Symfony\Component\Serializer\Mapping\Factory\ClassMetadataFactory;
+use Symfony\Component\Serializer\Mapping\Loader\AttributeLoader;
+use Symfony\Component\Serializer\Mapping\Loader\LoaderChain;
+use Symfony\Component\Serializer\Mapping\Loader\LoaderInterface;
+use Symfony\Component\Serializer\Mapping\Loader\XmlFileLoader;
+use Symfony\Component\Serializer\Mapping\Loader\YamlFileLoader;
+
+/**
+ * Warms up serializer metadata.
+ *
+ * @author Titouan Galopin <galopintitouan@gmail.com>
+ */
+final class SerializerCacheWarmer extends AbstractPhpFileCacheWarmer
+{
+    /**
+     * @param LoaderInterface[] $loaders      The serializer metadata loaders
+     * @param string            $phpArrayFile The PHP file where metadata are cached
+     */
+    public function __construct(
+        private array $loaders,
+        string $phpArrayFile,
+    ) {
+        parent::__construct($phpArrayFile);
+    }
+
+    protected function doWarmUp(string $cacheDir, ArrayAdapter $arrayAdapter, ?string $buildDir = null): bool
+    {
+        if (!$buildDir) {
+            return false;
+        }
+        if (!$this->loaders) {
+            return true;
+        }
+
+        $metadataFactory = new CacheClassMetadataFactory(new ClassMetadataFactory(new LoaderChain($this->loaders)), $arrayAdapter);
+
+        foreach ($this->extractSupportedLoaders($this->loaders) as $loader) {
+            foreach ($loader->getMappedClasses() as $mappedClass) {
+                try {
+                    $metadataFactory->getMetadataFor($mappedClass);
+                } catch (\Exception $e) {
+                    $this->ignoreAutoloadException($mappedClass, $e);
+                }
+            }
+        }
+
+        return true;
+    }
+
+    /**
+     * @param LoaderInterface[] $loaders
+     *
+     * @return list<XmlFileLoader|YamlFileLoader|AttributeLoader>
+     */
+    private function extractSupportedLoaders(array $loaders): array
+    {
+        $supportedLoaders = [];
+
+        foreach ($loaders as $loader) {
+            if (method_exists($loader, 'getMappedClasses')) {
+                $supportedLoaders[] = $loader;
+            } elseif ($loader instanceof LoaderChain) {
+                $supportedLoaders = array_merge($supportedLoaders, $this->extractSupportedLoaders($loader->getLoaders()));
+            }
+        }
+
+        return $supportedLoaders;
+    }
+}
