+++ src/Symfony/Component/Serializer/Tests/Fixtures/Serialization/Author.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Symfony\Component\Serializer\Tests\Fixtures\Serialization;
+
+class Author
+{
+    public $eyeColor;
+}
