+++ src/Symfony/Component/Serializer/Tests/Fixtures/Serialization/Person.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Symfony\Component\Serializer\Tests\Fixtures\Serialization;
+
+class Person
+{
+    public $eyeColor;
+}
