+++ src/Symfony/Component/Serializer/Tests/Fixtures/Serialization/Resources/author.yml
@@ -1,4 +1,4 @@
-Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Serialization\Author:
+Symfony\Component\Serializer\Tests\Fixtures\Serialization\Author:
     attributes:
         eyeColor:
             groups: ['group1', 'group2']
