+++ src/Symfony/Component/Serializer/Tests/Fixtures/Serialization/Resources/does_not_exist.yaml
