+++ src/Symfony/Component/Serializer/Tests/Fixtures/Serialization/Resources/person.xml
@@ -4,7 +4,7 @@
             xsi:schemaLocation="http://symfony.com/schema/dic/serializer-mapping
         https://symfony.com/schema/dic/serializer-mapping/serializer-mapping-1.0.xsd"
 >
-    <class name="Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Serialization\Person">
+    <class name="Symfony\Component\Serializer\Tests\Fixtures\Serialization\Person">
         <attribute name="eyeColor">
             <group>group1</group>
             <group>group2</group>
