+++ src/Symfony/Component/Serializer/composer.json
@@ -49,6 +49,7 @@
     "conflict": {
         "phpdocumentor/reflection-docblock": "<5.2|>=7",
         "phpdocumentor/type-resolver": "<1.5.1",
+        "symfony/cache": "<8.2",
         "symfony/dependency-injection": "<8.2",
         "symfony/property-access": "<8.1",
         "symfony/property-info": "<7.4.15",
