+++ src/Symfony/Component/Validator/Tests/Fixtures/Validation/Article.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Validation;
+namespace Symfony\Component\Validator\Tests\Fixtures\Validation;
 
 class Article implements NotExistingInterface
 {
