+++ src/Symfony/Component/Validator/Tests/Fixtures/Validation/Author.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Symfony\Component\Validator\Tests\Fixtures\Validation;
+
+class Author
+{
+    public $eyeColor;
+}
