+++ src/Symfony/Component/Validator/Tests/Fixtures/Validation/Resources/author.yml
@@ -1,4 +1,4 @@
-Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Validation\Author:
+Symfony\Component\Validator\Tests\Fixtures\Validation\Author:
     properties:
         eyeColor:
             - Choice: { choices: [brown, green, blue], message: Choose a valid eye color. }
