+++ src/Symfony/Component/Validator/Tests/Fixtures/Validation/Resources/categories.yml
@@ -1,10 +1,10 @@
-Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Validation\Category:
+Symfony\Component\Validator\Tests\Fixtures\Validation\Category:
     properties:
         id:
             - Type:
                 type: int
 
-Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Validation\SubCategory:
+Symfony\Component\Validator\Tests\Fixtures\Validation\SubCategory:
     properties:
         id:
             - Type:
