+++ src/Symfony/Component/Validator/composer.json
@@ -47,6 +47,7 @@
     },
     "conflict": {
         "doctrine/lexer": "<1.1",
+        "symfony/cache": "<8.2",
         "symfony/dependency-injection": "<8.2",
         "symfony/doctrine-bridge": "<7.4",
         "symfony/expression-language": "<7.4"
